requirements.txt
asyncpg==0.30.0
coverage==7.8.2
dotenv==0.9.9
filelock==3.18.0
fsspec==2025.5.1
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pluggy==1.6.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
typing_extensions==4.13.2
tzdata==2025.2
